pyproject.toml
[project]
name = "strands-agent"
version = "0.1.0"
description = "Local AI agent using Strands with Ollama"
requires-python = ">=3.12"

dependencies = [
    "strands-agents[a2a,ollama,openai]==1.3.0",
    "strands-agents-tools[local_chromium_browser]==0.2.3",
    "strands-agents-tools[mcp]==0.2.3",
    "strands-agents-tools==0.2.3",
    "python-dotenv",
    "SpeechRecognition",
    "PyAudio",
    "mcp>=1.12.3",
    "requests>=2.32.4",
]
